worker/requirements.txt
smbprotocol
pypdf
docx2txt
langchain-text-splitters
litellm
qdrant-client
python-dotenv
cryptography
